+++ src/Request/People.php
@@ -1746,15 +1746,13 @@ class People extends RequestCollection
         $nullState = false)
     {
         $views = [
-            'reshare_share_sheet',
-            'direct_user_search_keypressed',
             'story_share_sheet',
             'direct_user_search_nullstate',
-            'direct_inbox_active_now',
             'forwarding_recipient_sheet',
+            'direct_inbox_active_now',
             'call_recipients',
-            'reels_share_sheet',
-            'direct_ibc_inbox_discovery',
+            'reshare_share_sheet',
+            'direct_user_search_keypressed',
         ];
 
         if ($this->ig->isExperimentEnabled('56721', 0, false) !== true) {
@@ -1763,7 +1761,7 @@ class People extends RequestCollection
 
         $request = $this->ig->request('banyan/banyan/')
             ->addParam('is_private_share', false)
-            ->addParam('views', ($nullState) ? '["direct_ibc_nullstate"]' : json_encode(array_values($views)));
+            ->addParam('views', ($nullState === null) ? '["direct_ibc_nullstate"]' : json_encode(array_values($views)));
 
         if ($this->ig->isExperimentEnabled('54280', 0, false)) {
             $request->addParam('IBCShareSheetParams', json_encode(['size' => max($this->ig->getExperimentParam('54280', 2, 5), (int) $this->ig->getExperimentParam('54280', 5, 3))]));